chore: dissolve dimos/memory/ — relocate timeseries, drop dead code
dimos/memory/ was a misnamed util library plus dead code, colliding with
the live memory2/. Relocate the two live pieces and delete the rest:

- timeseries base + InMemoryStore -> dimos/utils/timeseries/
- LegacyPickleStore -> dimos/utils/testing/legacy_pickle.py (beside the
  replay/moment fixtures that wrap it)
- delete embedding.py (dead POC superseded by memory2; orphaned blueprint),
  the unused pickledir/sqlite/postgres timeseries backends, and the
  orphaned detect.py / fake_zed_module.py modules
- regenerate all_blueprints.py (drops the embedding-memory entry)
- rewrite all import sites + mock.patch targets; trim test_base.py to the
  InMemoryStore case

memory2/ untouched.
