+++ dimos/memory2/test_e2e.py
@@ -21,7 +21,6 @@ from typing import TYPE_CHECKING, Any
 
 import pytest
 
-from dimos.memory.timeseries.legacy import LegacyPickleStore
 from dimos.memory2.embed import EmbedImages
 from dimos.memory2.store.sqlite import SqliteStore
 from dimos.memory2.transform import QualityWindow
@@ -30,6 +29,7 @@ from dimos.msgs.sensor_msgs.Image import Image
 from dimos.msgs.sensor_msgs.PointCloud2 import PointCloud2
 from dimos.robot.unitree.type.odometry import Odometry
 from dimos.utils.data import get_data_dir
+from dimos.utils.testing.legacy_pickle import LegacyPickleStore
 
 if TYPE_CHECKING:
     from collections.abc import Iterator
