+++ dimos/robot/unitree/testing/test_tooling.py
@@ -16,10 +16,10 @@ import time
 
 import pytest
 
-from dimos.memory.timeseries.legacy import LegacyPickleStore
 from dimos.robot.unitree.type.lidar import pointcloud2_from_webrtc_lidar
 from dimos.robot.unitree.type.odometry import Odometry
 from dimos.utils.reactive import backpressure
+from dimos.utils.testing.legacy_pickle import LegacyPickleStore
 
 
 @pytest.mark.tool
