+++ dimos/types/test_timestamped.py
@@ -19,8 +19,6 @@ import pytest
 from reactivex import operators as ops
 from reactivex.scheduler import ThreadPoolScheduler
 
-from dimos.memory.timeseries.inmemory import InMemoryStore
-from dimos.memory.timeseries.legacy import LegacyPickleStore
 from dimos.msgs.sensor_msgs.Image import Image
 from dimos.types.timestamped import (
     Timestamped,
@@ -31,6 +29,8 @@ from dimos.types.timestamped import (
 )
 from dimos.utils.data import get_data
 from dimos.utils.reactive import backpressure
+from dimos.utils.testing.legacy_pickle import LegacyPickleStore
+from dimos.utils.timeseries.inmemory import InMemoryStore
 
 
 def test_timestamped_dt_method() -> None:
