+++ dimos/utils/testing/test_replay.py
@@ -17,11 +17,11 @@ import re
 import pytest
 from reactivex import operators as ops
 
-from dimos.memory.timeseries.legacy import LegacyPickleStore
 from dimos.msgs.sensor_msgs.PointCloud2 import PointCloud2
 from dimos.robot.unitree.type.lidar import pointcloud2_from_webrtc_lidar
 from dimos.robot.unitree.type.odometry import Odometry
 from dimos.utils.data import get_data
+from dimos.utils.testing.legacy_pickle import LegacyPickleStore
 
 pytestmark = pytest.mark.self_hosted
