+++ dimos/utils/timeseries/base.py
