+++ dimos/utils/timeseries/test_base.py
@@ -14,19 +14,14 @@
 """Tests for TimeSeriesStore implementations."""
 
 from dataclasses import dataclass
-from pathlib import Path
 import tempfile
-import uuid
 
 import pytest
 from reactivex import operators as ops
 
-from dimos.memory.timeseries.base import TimeSeriesStore
-from dimos.memory.timeseries.inmemory import InMemoryStore
-from dimos.memory.timeseries.legacy import LegacyPickleStore
-from dimos.memory.timeseries.pickledir import PickleDirStore
-from dimos.memory.timeseries.sqlite import SqliteStore
 from dimos.types.timestamped import Timestamped
+from dimos.utils.timeseries.base import TimeSeriesStore
+from dimos.utils.timeseries.inmemory import InMemoryStore
 
 
 @dataclass
@@ -56,66 +51,9 @@ def make_in_memory_store() -> TimeSeriesStore[SampleData]:
     return InMemoryStore[SampleData]()
 
 
-def make_pickle_dir_store(tmpdir: str) -> TimeSeriesStore[SampleData]:
-    return PickleDirStore[SampleData](tmpdir)
-
-
-def make_sqlite_store(tmpdir: str) -> TimeSeriesStore[SampleData]:
-    return SqliteStore[SampleData](Path(tmpdir) / "test.db")
-
-
-def make_legacy_pickle_store(tmpdir: str) -> TimeSeriesStore[SampleData]:
-    return LegacyPickleStore[SampleData](Path(tmpdir) / "legacy")
-
-
-# Base test data (always available)
 testdata: list[tuple[object, str]] = [
     (lambda _: make_in_memory_store(), "InMemoryStore"),
-    (lambda tmpdir: make_pickle_dir_store(tmpdir), "PickleDirStore"),
-    (lambda tmpdir: make_sqlite_store(tmpdir), "SqliteStore"),
-    (lambda tmpdir: make_legacy_pickle_store(tmpdir), "LegacyPickleStore"),
 ]
-
-# Track postgres tables to clean up
-_postgres_tables: list[str] = []
-
-try:
-    import psycopg2
-
-    from dimos.memory.timeseries.postgres import PostgresStore
-
-    # Test connection
-    _test_conn = psycopg2.connect(dbname="dimensional")
-    _test_conn.close()
-
-    def make_postgres_store(_tmpdir: str) -> TimeSeriesStore[SampleData]:
-        """Create PostgresStore with unique table name."""
-        table = f"test_{uuid.uuid4().hex[:8]}"
-        _postgres_tables.append(table)
-        store = PostgresStore[SampleData](table)
-        store.start()
-        return store
-
-    testdata.append((lambda tmpdir: make_postgres_store(tmpdir), "PostgresStore"))
-
-    @pytest.fixture(autouse=True)
-    def cleanup_postgres_tables():
-        """Clean up postgres test tables after each test."""
-        yield
-        if _postgres_tables:
-            try:
-                conn = psycopg2.connect(dbname="dimensional")
-                conn.autocommit = True
-                with conn.cursor() as cur:
-                    for table in _postgres_tables:
-                        cur.execute(f"DROP TABLE IF EXISTS {table}")
-                conn.close()
-            except Exception:
-                pass  # Ignore cleanup errors
-            _postgres_tables.clear()
-
-except Exception:
-    print("PostgreSQL not available")
 
 
 @pytest.mark.parametrize("store_factory,store_name", testdata)
